requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
sentence-transformers==2.2.2
numpy==1.26.3
groq==0.4.2
google-generativeai==0.3.2